requirements.txt
functions-framework==3.8.2
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.85.0
google-genai==1.8.0
pydantic
requests>=2.31.0
python-docx==1.1.2 
PyMuPDF
google-auth==2.38.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
tenacity==8.2.2
google-cloud-firestore==2.20.1
pybreaker==1.3.0
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-instrumentation-flask
opentelemetry-exporter-gcp-trace==1.8.0
opentelemetry-propagator-gcp==1.8.0
PyJWT>=2.8.0
cryptography>=3.0.0
google-cloud-secret-manager==2.23.2
google-adk>=0.1.0
supabase>=1.2.0
python-dotenv>=1.0.0
httpx>=0.28.1,<1.0.0
argparse>=1.4.0
vertexai